marlin/sailfish: Add libartbase and libziparchive to VNDK-SP
They are new indirect dependencies of libRS_internal.so.
The dependency is cut for VNDK-enabled devices, but is still alive for
semi-VNDK-enabled devices like marlin/sailfish.

Put the two libs to the list so that they are copied to /system/lib/vndk-sp/.

BenzoEdit: Add since we use some art commits from master and have libartbase

Bug: 110478183
Test: run any app using renderscript
Change-Id: I434399e1dbaba7399a9cbefb089b3dccb8897bee
Signed-off-by: mydongistiny <jaysonedson@gmail.com>